plot: refactored into packages
