+++ src/main/java/io/github/mzmine/gui/chartbasics/simplechart/datasets/ColoredXYDataset.java
@@ -1,30 +1,31 @@
 /*
- * Copyright 2006-2020 The MZmine Development Team
+ *  Copyright 2006-2020 The MZmine Development Team
  *
- * This file is part of MZmine.
+ *  This file is part of MZmine.
  *
- * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
+ *  MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
+ *  General Public License as published by the Free Software Foundation; either version 2 of the
+ *  License, or (at your option) any later version.
  *
- * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
+ *  MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ *  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ *  Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
- * USA
+ *  You should have received a copy of the GNU General Public License along with MZmine; if not,
+ *  write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+ *  USA
  */
 
-package io.github.mzmine.gui.chartbasics.template;
+package io.github.mzmine.gui.chartbasics.simplechart.datasets;
 
-import io.github.mzmine.gui.chartbasics.template.providers.ColorPropertyProvider;
-import io.github.mzmine.gui.chartbasics.template.providers.ColorProvider;
-import io.github.mzmine.gui.chartbasics.template.providers.LabelTextProvider;
-import io.github.mzmine.gui.chartbasics.template.providers.PlotXYDatasetProvider;
-import io.github.mzmine.gui.chartbasics.template.providers.SeriesKeyProvider;
-import io.github.mzmine.gui.chartbasics.template.providers.ToolTipTextProvider;
-import io.github.mzmine.gui.chartbasics.template.providers.XYValueProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.SimpleXYLineChart;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.ColorPropertyProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.ColorProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.LabelTextProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.PlotXYDatasetProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.SeriesKeyProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.ToolTipTextProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.XYValueProvider;
 import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.taskcontrol.Task;
 import io.github.mzmine.taskcontrol.TaskPriority;
