+++ src/main/java/io/github/mzmine/gui/chartbasics/simplechart/datasets/ColoredXYZDataset.java
@@ -1,0 +1,64 @@
+/*
+ *  Copyright 2006-2020 The MZmine Development Team
+ *
+ *  This file is part of MZmine.
+ *
+ *  MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
+ *  General Public License as published by the Free Software Foundation; either version 2 of the
+ *  License, or (at your option) any later version.
+ *
+ *  MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ *  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ *  Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with MZmine; if not,
+ *  write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+ *  USA
+ */
+
+package io.github.mzmine.gui.chartbasics.simplechart.datasets;
+
+import io.github.mzmine.gui.chartbasics.simplechart.providers.ColorProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.LabelTextProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.SeriesKeyProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.ToolTipTextProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.XYValueProvider;
+import org.jfree.data.xy.XYZDataset;
+
+public class ColoredXYZDataset extends ColoredXYDataset implements XYZDataset {
+
+  public ColoredXYZDataset(
+      XYValueProvider XYValueProvider,
+      SeriesKeyProvider<Comparable<?>> seriesKeyProvider,
+      LabelTextProvider labelTextProvider,
+      ToolTipTextProvider toolTipTextProvider,
+      ColorProvider colorProvider) {
+    super(XYValueProvider, seriesKeyProvider, labelTextProvider,
+        toolTipTextProvider, colorProvider);
+  }
+
+
+  /**
+   * Returns the z-value for the specified series and item.
+   *
+   * @param series the series index (zero-based).
+   * @param item   the item index (zero-based).
+   * @return The z-value (possibly {@code null}).
+   */
+  @Override
+  public Number getZ(int series, int item) {
+    return null;
+  }
+
+  /**
+   * Returns the z-value (as a double primitive) for an item within a series.
+   *
+   * @param series the series (zero-based index).
+   * @param item   the item (zero-based index).
+   * @return The z-value.
+   */
+  @Override
+  public double getZValue(int series, int item) {
+    return 0;
+  }
+}
